Updated openstack/openstack Project: openstack-infra/project-config  b19c9c1e54a446196bcca86e737b69cb8400b4bc
Remove experimental jobs *-with-oslo-master

In Ibdfc03f27450a5392acc276f98bfb464f9a0f663, we added
some periodic jobs to the the experimental queue so we
could trigger them on demand to try out the jobs. Now
that the periodic jobs are working fine, we should clean
up the experimental queue.

Change-Id: I7408e700e3e63ff76f8e19f999ebfdc02882395d