Updated openstack/openstack Project: openstack-infra/project-config  a613147c251c7d9762880533233c211d7c622526
Move Neutron Grenade multi from experimental to check queue

During testing, an issue with Nova was identified and resolved
in change f2c1cfae. This addresses the persistent errors seen
on the Grenade job.

Now that we can get successful runs, the job is ready to move
to the check queue.

Depends-on: I22eb3a3fcd8e74a1d9085acde15c25a927ae12cb
Change-Id: Icfaaf075dda3ecc89a16f65962cc6c673fb7e3ae