Updated openstack/openstack Project: openstack-infra/project-config  1460e208e8122f2b70fd33e48b0ae9a5d452df53
Fix rally skip rules

Changes like [1] revealed that we run rally jobs even
if we do not have to. It turns out a skip-if clause
added to the rally project masked a precendent one,
thus triggering rally runs on any doc change for any
project except Rally.

Move the skip rules to the proper place so that rally jobs are skipped
as intented.

This skips for rally also the gate-rally-dsvm-cli job which was run
before. The rally file match list has been adjusted to remove the
samples directory, jobs should be run if it gets updated.

[1] https://review.openstack.org/#/c/281628/

Change-Id: I506d1d4fb479b7a2216f5c3539b23442443ec28d