Updated openstack/openstack Project: openstack/nova  437dbebc1f3b6f95d51cca33240b06396a30951d
Add better help text to scheduler options (7)

The previous commit moved all the scheduler options to a central
location. This commit adds additional help text to those options.
Because of the large number of options, the improved help text will be
split into several patches; this is patch 7 of 7.

Blueprint centralize-config-options

Change-Id: Id48c9a9b59b2f57531edc7983522e6aa6b4f13bd