Updated openstack/openstack Project: openstack/openstack-manuals  44a47c86d879ddbd3c8e3d7e7d44711a1bb0cb8f
Setting up Administrator Guide Dashboard Chapter

Introducing Admin User Guide dashboard content to
the Cloud Admin Guide as a part of the reorganisation
goal. This patch is the first step in creating a new
Dashboard section for Admin Users in the Cloud Admin Guide,
as disucssed in the User Guide Specialty team meetings.

1.) Moving:
   dashboard_manage_host_aggregates.rst
   dashboard_manage_flavors.rst
   dashboard_admin_manage_stacks.rst
   dashboard_manage_instances.rst
   dashboard_manage_images.rst
   shared_file_systems_manage_shares_dashboard.rst

2). Remove the Images and Instances content from the compute-images-instances.rst

3.) Move the Shared file system dashboard content out of
    shared_file_system.rst, and into the Dashboard.rst chapter.

Change-Id: I1e3c122e58349853b61be4ba514e469da407c1c9
Implements: blueprint user-guides-reorganised