Updated openstack/openstack Project: openstack/taskflow  4389c1a9d878d8422c49d74d1da69f32eb1dff07
Sqlalchemy-utils double entry (already in test-requirements.txt)

Since sqlalchemy as a backend is optional (only needed if someone
is using persistence and a database) then it's only needed during
testing (to ensure it still works).

Also pip complains about double requirement entries and fails to
install into a virtualenv, so this fixes that as well.

Change-Id: Ica34b719e7a750a9cf3ae381f43edf117e0478a8