Updated openstack/openstack Project: openstack/nova  ce7bfcc350f23552ce6e057e37542292c7ae0b3f
Update HostManager and DiskFilter to use ComputeNode disk ratio

Since the ComputeNode object provides a compatible facade for the Scheduler,
we can get it from the HostManager and provide this field up to the filter.

As the HostManager is calling ComputeNodeList.get_all(), then the related HostStates
will all have the scheduler nova.conf allocation ratios.

Change-Id: Idf0202d53239862c208ad2d68167c30086b6024f
Partially-Implements: blueprint disk-allocation-ratio-to-rt