Updated openstack/openstack Project: openstack/nova  dc752eb6a210402db359ca8d9de3f30bb7354176
Update ComputeNode values with disk allocation ratios in the RT

Now that we have added the field for persisting the disk alloc ratio, we can
have the ResouceTracker persisting it by adding it to the local ComputeNode
object which is persisted by calling the _update() method.
It will then send by default 0.0 unless the operator explicitely specified an
allocation ratio in the compute nova.conf.

Thanks to the ComputeNode object hydratation on the scheduler side, the facade
will make sure that if a default 0.0 is provided by either a compute node or
by the scheduler's nova.conf, it will actually get the original allocation
ratios (ie. 1.0 for disk)
Since the Scheduler reads the same RT opt but goes thru the ComputeNode object,
it will also get the Facade returning 1.0 unless the operator
explicitely provided other ratios for the scheduler's nova.conf

Amending the release note now that the behaviour is changing.

DocImpact Disk alloc ratio is now per computenode
UpgradeImpact

Change-Id: Ief6fa32429d58b80e70029ed67c7f42e0bdc986d
Implements: blueprint disk-allocation-ratio-to-rt