Updated openstack/openstack Project: openstack/nova  8aa60ba501e87b5c2f22ad1494cc3bcb2859161e
Update example policy.json to remove "" policies

A blank policy is equivalent to a policy that allows everything. In
practice this is probably not what you want, and there is no indication
that the policies were set this way for that reason, except for one
case. It appears to be a matter of copy/paste.

The exception is the policy check for default quota-sets. The default
quota set is not scoped to a project or user so checking for ownership
does not make sense. And since it is the same for every project there's
no reason to require an admin context in order to view it. So it's set
to "@" here to be visible to all.

The rest of the calls end up having ownership verified in the db api by
filtering on project_id, unless it's an admin context. So admin_or_owner
ends up being equivalent to what's in place, but verification happens at
the policy level which is where it should be checked.

Change-Id: I698c8aa54e90fc285a4a6e0a9105e8d1ccb90c8a