Updated openstack/openstack Project: openstack/nova  9386801ce2d1406053d44eb79f8423b02ce04498
Fixed leaked UnexpectedMethodCallErrors in test_compute

Both of these tests were actually leaking UnexpectedMethodCallErrors
from mox due to the allocate_for_instance call being stubbed out
but the kwargs were in the wrong order.

Apparently something with mox, the stub_out fixture, and
NetworkInfoAsyncWrapper using nova.utils.spawn cause the stubs
to be on a greenthread so when mox raises the error, it doesn't
fail the actual test, it just gets logged to stderr by eventlet.

In the case of the test_run_instance_queries_macs test, this was
actually causing the test to run over a minute (presumably because
it would eventually hit some timeout in the thread, like an rpc
timeout maybe?). With this fix it drops that down to around 1 second
to run the test.

We really need to figure out what's causing the stub to go on the
thread so that the test doesn't get the UnexpectedMethodCallError
and fail but that could come in a later change if someone can
figure it out.

Change-Id: Ibc5c881e17310304eb63f5df6360f8cb1657b807
Closes-Bug: #1547279