Updated openstack/openstack Project: openstack/designate  602a74797296bc97ec641dd40545ca683d3e9aa1
Improve performance and utility of Recovery

This makes a couple small, but significant changes to periodic recovery.

- Change the CUD calls for periodic recovery to use the Pool Manager's RPCAPI.

Periodic recovery is charged with fixing problems that have come about in
the normal course of operation. Sometimes things go very wrong and there
are a lot of resources that have an ERROR state. This change allows the work
generated to be distributed among active `designate-pool-manager` processes.
Hopefully resolving the issues faster, and reducing risk of random failure on
the one process generating the recovery events.

- Include zones that have been PENDING for a long time

After a certain time, a product of timeouts, intervals, and delays, changes
either should have successfully propagated, or gone to an ERROR state.
Unfortunately this doesn't happen on occaison. Certain random failures and
undiscovered bugs leave zones hanging out in PENDING state forever. By treating
these as states as errors, we (hopefully) ensure that Designate will clean up
these zones and return them to the happy path.

Change-Id: Ifb375bc9383e7dd8fd6e51a42a4a50dc5015cb13