Updated openstack/openstack Project: openstack/swift  3e3ad5cb160a14c29afdf6fee8ac4a43f8d8416f
Fail early if the memcache address is invalid.

In cases when the memcache address is invalid, we should fail early.
This patch addresses the cases when the IPv6 addresses are not
enclosed in "[]". It does not, however, fix the case of an invalid
hostname. These improvements could also be added to the _get_addr()
method.

Change-Id: I4743dcda45a1fc1640989325c4a2e1fea591fc69