Updated openstack/openstack Project: openstack/neutron-lbaas  eb52dea0700b8be3086dd2ad7333125a32befe77
Shared pools support

In preparation for L7 switching functionality, we need to
reduce the rigidity of our model somewhat and allow pools
to exist independent of listeners and be shared by 0 or
more listeners. With this patch, pools are now associated
with loadbalancers directly, and there is now a N:M
relationship between listeners and pools.

This patch does alter the Neutron LBaaS v2 API slightly,
but all these changes are backward compatible. Nevertheless,
since Neutron core dev team has asked that any API changes
take place in an extension, that is what is being done in
this patch.

This patch also updates the reference namespace driver to
render haproxy config templates correctly given the pool
sharing functionality added with the patch.

Finally, the nature of shared pools means that the usual
workflow for tenants can be (but doesn't have to be)
altered such that pools can be created before listeners
independently, and assigned to listeners as a later step.

DocImpact
APIImpact
Partially-Implements: blueprint lbaas-l7-rules
Change-Id: Ia0974b01f1f02771dda545c4cfb5ff428a9327b4