Updated openstack/openstack Project: openstack/barbican  c06fb13ff912b2a2570a27e9a8ba070507116cde
Fix http 500 when getting secret payload with no Acccept header

Functional tests had default values for Accept header which were
masking a bug when you try to GET a secret with /payload and don't
pass in an accept header.

Fixed this and also updated the functional test client to allow
tests to specify headers to be omitted on a test-by-test basis.

Change-Id: Ia67d37571ca3d561cdcc67ce3c4fb896def35a24