Updated openstack/openstack Project: openstack/neutron  29ccfeebf775391cb1ca56eee00fa1f9987099a7
Revert "tests: Collect info on failure of conn_tester"

More info in bug report. I suggest we first revert, then
re-introduce the collect_debug_info patch with a different
approach. I suspect the fix is not trivial if indeed ordering
is an issue and the namespaces are cleaned up before
collect_debug_info is fired.

Related-Bug: #1548547
Change-Id: Ice93abbc6e143cdbb90e7d41d1be86dc9eb05006