Updated openstack/openstack Project: openstack/tooz  af9a4a6c29d5b86b9362402e83c13d76c13dbd22
Fix calling acquire(blocking=False) twice leads to a deadlock

This patch is fixing a possible deadlock caused when
acquire(blocking=False) is invoked more than once. The zookeeper, ipc,
memcached and redis drivers had problems with this.

For zookeeper, memcached and redis the fix consist in avoiding overwriting
the value of the "acquired" flag every time acquired() is invoked,
instead we just rely on the backend library to tell us whether the lock
has been acquired or not.

For ipc the fix is in two places: (a) when acquire() is invoked we
need to make sure we overwrite the "self._lock" attribute only when we
actually acquired the lock, not in failure cases; for that we are using
a temporary variable to hold the lock until its successfully acquired.
(b) when release() is invoked we should only set the "self._lock"
attribute to None if we actually released the lock successfully.

Closes-Bug: #1543688
Change-Id: I652e168028f222230cb87c7a8f42bda02b4dd1f2