Updated openstack/openstack Project: openstack/neutron  d8238e5b536e6ed384e8c0a6a31b5da11893543c
stadium: Add a guideline about contributor overlap

This patch introduces a criterion whereby sub-projects are considered
for inclusion (or exclusion) under the Neutron fold based on team
overlap, i.e. on who is driving the development of the sub-project
from inception to production.

For instance, Neutron team members may occasionally need to create
new repos for experimentation, or for the development of new
(potentially overlapping) functional areas to be incubated as an
integral part of the Neutron end-to-end system.

This patch is the last in a series that introduces criteria for
classification of projects for inclusion/exclusion to/out of the
Neutron 'stadium'. These criteria will always be applied jointly
when assessing a project for inclusion/exclusion at discretion of
the Neutron team.

Patches that complete the inclusion/exclusion process will follow
to discuss the rationale for inclusion/exclusion of the remainder
of the sub-projects.

Change-Id: Ie8e65a94ff9ec238e7c28a8b85d92835675d0c4c
Signed-off-by: Russell Bryant <rbryant@redhat.com>
Co-authored-by: Armando Migliaccio <armamig@gmail.com>