Updated openstack/openstack Project: openstack-infra/project-config  6b71e8cac676e04141839eeecce3462df3a04848
Remove check-deletions from openstack-manuals

The check-deletions job is obsolete on this repository, remove it.
The repository is mainly RST (with one exception) and thus this XML
specific check is not needed.

It's part of the template openstack-doc-gate, so use openstack-rst-gate
instead and add those jobs that are not in the rst-gate.

Change-Id: I601ee6308c46a2ab322b2ae6160ebc4692b2634f