Updated openstack/openstack Project: openstack/requirements  3df185477fc56c4c98aa369eb18c7f3bbab0e6b2
Adds PyMI as a global-requirement

PyMI is a drop-in replacement for the old, unmaintained WMI module.
The new module maintains the same interface, while providing better
error handling, better multithreading support, better performance
and a higher reliability.

Maintenance. PyMI is developed by Cloudbase Solutions, active
contributors to the OpenStack project. Any bug found has been dealt
with as soon as possible. Also, it does not depend on the pywin32
library. Recent history can be observed at [1]. Meanwhile, the latest
WMI release was on 2011.04.13. [2]

Code quality. The code has been tested repeatedly on the Hyper-V CI and
using rally. Any and all found issues that occured during testing has
been solved. The performance gain was noticeable. Succesful Hyper-V CI
runs using PyMI can be observed at [3][4][5][6].

Python 3. PyMI is python 3 compatible, the python 3.4 wheel is available
on pypi [7].

License. The library is Apache 2.0 licensed [8].

Packaging. Wheels for python 2.7 and 3.4 can be found on pypi. This
library is meant for Windows operating systems.

Other libraries. As mentioned before, WMI has the same functionality as
PyMI. Compared to PyMI, WMI it is unmaintained and outdated, we've hit
plenty of issues and bugs using it, it is slow, it does not have any
multithreading support, and it also depends on pywin32.

Replacing WMI with PyMI can lead to an enhanced experience overall.

[1] https://github.com/cloudbase/PyMI/commits/master
[2] https://pypi.python.org/pypi/WMI/
[3] https://review.openstack.org/#/c/256611/
[4] https://review.openstack.org/#/c/256616/
[5] https://review.openstack.org/#/c/256614/
[6] https://review.openstack.org/#/c/256585/
[7] https://pypi.python.org/pypi/PyMI/1.0.0
[8] https://github.com/cloudbase/PyMI/blob/master/LICENSE

Change-Id: Ia2d059ed4550f0aaaea5d8da10a254ca67674eec