Updated openstack/openstack Project: openstack/governance  5c7e9440688f619f42438a3f7a2b2afd3e7c9819
Propose an update to the OpenStack mission statement.

This patch is a follow-up to our previous resolution suggesting an
update to the OpenStack mission statement.  The OpenStack Foundation
Board of Directors had feedback and requested it be discussed further on
the foundation mailing list.  This version reflects the outcome of that
discussion.

Change-Id: I16a649872414099b394ec1543fb6e73e7de72151