Updated openstack/openstack Project: openstack/governance  053f75008adea5b231a50efa788f4bcc6773165a
Add project Dragonflow to OpenStack big-tent

Following the directive
http://governance.openstack.org/reference/new-projects-requirements.html

OpenStack Mission Alignment:
Dragonflow is a distributed SDN controller for OpenStack Neutron
supporting distributed Switching, Routing, DHCP and more.
Our project mission is to implement advanced networking services in
a manner that is efficient, elegant and simple.

Dragonflow is designed to support large scale deployments with a
focus on latency and performance, as well as providing advanced
innovative services that run locally on each compute node,
with container technology in mind.

Our Mission Statment:

1) Implement Neutron APIs using SDN principles,
   while keeping both Plug-in and Implementation fully
   under OpenStack project and governance.

2) 100% open source, contributors are welcome to partner
   and share a mutual vision.

3) Lightweight and Simple in terms of code size and
   complexity, so new users / contributors have a simple
   and fast ramp-up.

4) Aim for performance-intensive environments, where
   latency is a big deal, while being small and
   intuitive enough to run on small ones as well.
   Completely pluggable design, easy to extend and enhance.

5) We *truly* believe in a distributed control plane.

Following the 4 opens
Open Source:
Dragonflow is 100% open source, everything from implementation
to design to future features and roadmap are happening and shared
openly and presented in various meetups and documents.
Dragonflow aims to become production grade open source OpenStack
Networking API implementation.

Open Community:
We are working closely with the community to share use cases and problems
and decide on future priorites for the project.
There are already few contributors from different companies which both
contribute code and both plan to deploy Dragonflow.
It is in our mission to collobrate with as many members/companies
in the community as possible and we welcome any feedback and any desire
to help us shape Dragonflow future by others.

Open Development:
- All Dragonflow code is being code reviewed in OpenStack CI
- Dragonflow has a core team which openly discuss all issues
- Dragonflow support gate tests which run unit tests/fullstack/tempest
  and Rally tests.
- Dragonflow collaborates with Neutron and put Neutron objects/data model
  as first class citizen in our design and implementation
- Dragonflow has no Northbound API other then the OpenStack/Neutron API
- Bugs are managed by OpenStack launchpad

Open Design:
- All designs and specs are published for review and managed in
  OpenStack launchpad
- Dragonflow conducts a weekly IRC meeting to discuss all designs
  and future roadmap
- Dragonflow repository consists of many Documentation and diagrams
  explaining the current design and code and also future roadmap
  and ideas.
- Everything is discussed openly either on the review board/
  Dragonflow IRC channel (which is logged) or the ML.
- Dragonflow mission statement is to become an integral part of OpenStack

Change-Id: I2e36c75baabef2a9e78e1c466abc5a1597b66ca4