Updated openstack/openstack Project: openstack/governance  5b362996c5e15856368d6f5990fe82b4c719dce6
Add project Kuryr to OpenStack big-tent

Following the directive
http://governance.openstack.org/reference/new-projects-requirements.html

OpenStack Mission Alignment:

OpenStack and Neutron are no longer the new kid in the block, Neutron
has matured and its popularity in OpenStack deployments over
nova-network is increasing, it has a very rich ecosystem of plugins
and drivers which provide networking solutions and services (like
LBaaS, VPNaaS and FWaaS).  All of which implement the Neutron
abstraction and intend to be interchangeable for cloud deployers.

What we noticed in regards to container networking, and specifically
in environments that are mixed for containers, virtual machines, and
bare metal instances is that every networking solution for virtual
machines and bare metal instances tries to reinvent and enable
networking for containers but this time using the Docker API (or any
other abstraction). OpenStack Magnum, for example, has to introduce an
abstraction layer for different libnetwork drivers depending on the
Container Orchestration Engine used.  It would be ideal if Kuryr could
be the default for Magnum COEs

The idea behind Kuryr is to be able to leverage the abstraction and
all the hard work that was put in Neutron and its plugins and services
and use that to provide production grade networking for containers use
cases. Instead of each independent Neutron plugin or solution trying
to find and close the gaps, we can concentrate the efforts and focus
in one spot - Kuryr.

Kuryr aims to be the “integration bridge” between several communities,
Containers communities (Docker / Kubernetes / Mesos) and Neutron and
propose and drive changes needed in Neutron (or in containers
frameworks) to be able to fulfill the use cases needed specifically in
container networking.

It is important to note that Kuryr is NOT a networking solution by
itself nor does it attempt to become one. The Kuryr effort is focused
to be the courier that delivers Neutron networking and services to
Docker / Kubernetes / Mesos.

Following the 4 opens

Open Source:
Kuryr is 100% open source, everything from implementation
to design to future features and roadmap are happening and shared
openly and presented in various meetups and documents.

Open Community:
Kuryr work flows and stats already speak for them selfs on this
front, Kuryr has a versatile community and is solving
use cases for many different companies while discussing, reviewing
and implementing them together

Open Development:
- All Kuryr code is being code reviewed in OpenStack CI
- Kuryr has a core team which openly discuss all issues and is made
  up with people from different companies
- Kuryr support gate tests which run unit tests/fullstack/tempest
  and Rally tests.
- Kuryr collaborates with Neutron, Magnum and Kolla and all the
  various Neutron backends and Neutron advance services

Open Design:
- All designs and specs are published for review and managed in
  OpenStack launchpad
- Kuryr conducts a weekly IRC meeting to discuss all designs
  and future roadmap.
  The IRC meeting is on alternating times to fit all different
  time zones
- Kuryr repository consists of many Documentation and diagrams
  explaining the current design and code and also future roadmap
  and ideas.
- Everything is discussed openly either on the review board/
  Kuryr IRC channel (which is logged) or the ML.

Change-Id: I81d4677f9d0081d1b967305956c1bec85cb765ea