Updated openstack/openstack Project: openstack/oslo.versionedobjects  83342c274fea7f6ba815b8bfc60942f437999c35
Fix messages in exceptions raised due to invalid state transitions

The recently-added FakeStateMachineField raises exceptions with strings
that have improper grammar, such as:

  FooObject's are not allowed transition out of ERROR state to ACTIVE ...

In addition to the grammar issue, the field's name in the object is not
provided which may make it very difficult to debug.

Finally, the tests for this field did not actually validate the string
of the exception.

This patch corrects the strings, includes the field name, and makes the
test actually validate that the output looks like we expect.

Change-Id: I7bee955fba15b9570416518b16ac64181bf7e320