Updated openstack/openstack Project: openstack/trove  7395cd9b0b4f2d61bbe00abb5d6fa9c5196189cb
Pass datastore details when required (again)

This is another set of fixes to integration tests, as the datastore is not
inherited by the parent job when a backup or a replica is created.

See also:
- Ib3e78bb68aaf992177b9d9cd1d89e84fb54830ed
- Ice626b3d3f73e75222b1080afa58232e03459a8e

Also, fix the fake guestagent so that it works with the increased
timing coming from the fixed (now working) tests.

Change-Id: I8d2b5a4627eb9292ba4535696f9565a3c20a0783