Updated openstack/openstack Project: openstack/nova  a7b9fecd54b907e0f42883bd31a97bf857352370
resource-provider versioned objects

Add ResourceProvider, Inventory and InventoryList objects with
minimal methods for persistence. It is likely the interfaces will
need to be expanded as these things are actually used.

An inventory can be created or updated or a list of inventories can
be retrieved by resource provider uuid.

A resource provider can be created and retrieved by uuid.

Inventory shares some code with a forthcoming Allocation object that
used to be in this patch will now come later. That sharing is via a
_HasAResoureProvider superclass.

Tests are mixed between unit and functional. The functional tests
explore the arc of an expected task.

Co-Authored-By: Dan Smith <dansmith@redhat.com>
Partially-Implements: blueprint resource-providers
Change-Id: I578eebf04c3b5d881aa7be7b72691515f2cf6758