Updated openstack/openstack Project: openstack/ceilometer  e397e9a519480fe6776f3cb45cd064d9f9093a74
Initial seed of hacking

It is an initial seed of hacking rules based on neutron and
nova projects.
For ceilometer it prevents bug 1508442 and closes/prevents
bug 1529836.

All the requirements about the hacking rules were corrected
on this propose review.
To be clear about the requirements, the only alarmed was
about import[1], in specific H306
[H306] "Alphabetically order your imports by the full module
path. Organize your imports"

[1] http://docs.openstack.org/developer/hacking/

Change-Id: Ia530873078c97925bfb91fde87b38131426a74c2
Closes-Bug: 1529836