Updated openstack/openstack Project: openstack/heat  7f41025fb77548798953f29b9bdab268977f3264
Fix resource definition loading in resource groups

When using template-validate, values do not need to be specified for
each parameter. However, when resource groups were used, an error was
raised for parameters in the resource-def that did not have a value.
This fix ignores missing parameter values when parsing through the
resource definition.

Change-Id: I0b455baa78ee53724678a4476063bb7429ec03f8
Partial-Bug: #1508857