Updated openstack/openstack Project: openstack/swift  090d7c9764b9c10296762971f52fd26abc7feb9d
Add warn method back to LogAdapter

The change in 684c4c04 to improve py3 support breaks older middleware
that consumes the SwiftLogger.

Good for Swift code to not to use warn because it's deprecated.

Bad for LogAdapter to not faithfully provide the current python logging
module API (deprecated or otherwise).

Change-Id: Ib0baf66b7f33f91562867076ff0526a287d8e5a6