Updated openstack/openstack Project: openstack/nova  54679e3decce14814a1d92f1cf32c865bab20ae4
Add note about using OS-EXT-* prefix for attribute naming

The 2.3 microversion used the OS-EXT-SRV-ATTR prefix for
compat with v2.0 for the ec2 API split out of nova.

The 2.16 microversion added the 'host_status' extended
server attribute and doesn't use the prefix.

This adds a note that nothing after the 2.3 microversion
should be using the OS-EXT-SRV-ATTR prefix since it's an
artifact of v2.0 and extensions which are deprecated in
v2.1.

Similar notes are added for other extensions that add
extended attribute names to other resources like AZs,
volumes and images.

Change-Id: I60a53e13121b7280a41c3889ba86edcd2ae7bf83