Updated openstack/openstack Project: openstack/heat  ece9438802cbcf4641f5fbd9e11fe52cc482b47c
Fix scheduler compatibility with Python 3.7 harder

Always catch StopIteration, rather than allowing it to bubble up a level.
This is neccessary due to PEP0479, wherein it was decided that Python will
start converting this exception to a RuntimeError when it bubbles up,
beginning with Python 3.7.

This is a follow-up to f7edd0ba2d3354e57c106e20ad88c50a859db592, which
missed one case of where StopIteration could be raised.

Change-Id: I34bc32e7477bcaa0ddfe30179f1008b1be97c0a3