Updated openstack/openstack Project: openstack/tripleo-heat-templates  949b6d1cc247a66869c627a504436f1642481667
Allow vncproxy to work with ssl enabled

Right now our vncproxy settings are hard-coded to http and the
non-ssl port.  This change adds a vncproxy entry to the endpoint
map and uses those values to configure the proxy correctly on
compute nodes.  This is sufficient to get it working in my
environment with ssl enabled.

Change-Id: I9d69b088eef4700959b33c7e0eb44932949d7b71