Updated openstack/openstack Project: openstack/heat  825ba55b90f7d738a87dae6ddb3765a3b2895814
Ignore errors on old properties during update

When an update fails on a template resource, we still store the template
in the stack environment. Retrying the update will fail because of
missing parameters, but we can ignore those as they should be harmless
and only matter in this broken case.

Change-Id: I859ea72b1cc95162b6f498b6af65e4d9fdd9458f
Closes-Bug: #1543685