Updated openstack/openstack Project: openstack/ironic  8e206ff73556aa7079980cd851f8440a45a60b90
Add deprecated_for_removal config info in ironic.conf.sample

Configuration options has a parameter called "deprecated_for_removal" to
indicate that one option is planned to be removed in a future release.
This patch is extending the config generator script to add this
information to the output .sample file.

Change-Id: I3d3a69ea9ca700edf96ede947ff7b074e7978484
Closes-Bug: #1549351